+++ app/javascript/utilities/commonTypes.ts
@@ -76,8 +76,8 @@ export type TerrainType = "o" | "f" | "b" | "j" | "p" | "s" | "m" | "w" | "g" | 
   "x" | "c" | "v" | "y"
 export const terrainType: { [index: string]: TerrainType } = {
   Open: "o", Forest: "f", Brush: "b", Jungle: "j", Sand: "s", Marsh: "m", Water: "w",
-  Crater: "c", Cave: "v",
-  Grain: "g", Rough: "r", Orchard: "d", Soft: "t", Debris: "x", Shallow: "y", Palm: "p"
+  Crater: "c", Cave: "v", Grain: "g", Rough: "r", Orchard: "d", Soft: "t", Debris: "x",
+  Shallow: "y", Palm: "p"
 }
 
 export type RoadType = "t" | "d" | "p" | "a"