+++ app/models/scenarios/scenario_406.rb
@@ -7,53 +7,54 @@ module Scenarios
     ALLIES = ["usa"].freeze
     AXIS = ["jap"].freeze
     STATUS = "b"
-    VERSION = "0.5"
+    VERSION = "0.6"
 
     DATE = [1943, 11, 20].freeze
     LAYOUT = [15, 11, "x"].freeze
 
     ALLIED_UNITS = {
-      "0": {
-        list: [
-          :usa_leader_6_2,
-          :usa_leader_5_1,
-          [2, :usa_leader_4_1],
-          [10, :usa_marine_rifle_s],
-          [4, :usa_m1918_bar],
-          :usa_radio_8inch,
-          [2, :usa_lvt_1],
-          :usa_lvt_2,
-        ],
-      },
-      "2": {
-        list: [
-          :usa_leader_5_1,
-          [3, :usa_marine_rifle_s],
-          [2, :usa_m1918_bar],
-          :usa_lvt_1,
-          :usa_lvt_2,
-        ],
-      },
+      "0": { list: [
+        :usa_leader_6_2,
+        :usa_leader_5_1,
+        [2, :usa_leader_4_1],
+        [10, :usa_marine_rifle_s],
+        [4, :usa_m1918_bar],
+        :usa_radio_8inch,
+        :usa_lvt_1,
+        :usa_lvt_2,
+      ] },
+      "2": { list: [
+        [2, :usa_leader_5_1],
+        [4, :usa_marine_rifle_s],
+        :usa_m1918_bar,
+        :usa_m2_browning,
+        :usa_radio_8inch,
+        :usa_lvt_1,
+        :usa_lvt_2,
+      ] },
+      "4": { list: [
+        :usa_leader_5_1,
+        [4, :usa_marine_rifle_s],
+        [2, :usa_m1918_bar],
+        :usa_lvt_1,
+        :usa_lvt_2,
+      ] },
     }.freeze
 
     AXIS_UNITS = {
-      "0": {
-        list: [
-          [3, :jap_leader_4_1],
-          [12, :jap_snlf_s],
-          [2, :jap_crew_t],
-          [4, :jap_type_92_hmg],
-          [2, :jap_70mm_type_92],
-          [4, :bunker],
-        ],
-      },
-      "3": {
-        list: [
-          :jap_leader_4_1,
-          [4, :jap_snlf_s],
-          [2, :jap_type_99_lmg],
-        ],
-      },
+      "0": { list: [
+        [3, :jap_leader_4_1],
+        [12, :jap_snlf_s],
+        [2, :jap_crew_t],
+        [4, :jap_type_92_hmg],
+        [2, :jap_70mm_type_92],
+        [4, :bunker],
+      ] },
+      "3": { list: [
+        :jap_leader_4_1,
+        [4, :jap_snlf_s],
+        [2, :jap_type_99_lmg],
+      ] },
     }.freeze
 
     class << self
@@ -86,7 +87,6 @@ module Scenarios
           bunkers, pillboxes, and trenches, unleashed a relentless barrage,
           resulting in catastrophic Marine casualties within the first few
           hours of the assault.",
-
           "Despite the chaos and carnage, the Marines gradually pushed inland
           through intense, close-quarters combat.  Over the next three days,
           they engaged in a brutal battle against a determined enemy willing
@@ -112,15 +112,20 @@ module Scenarios
           allied_dir: 6,
           axis_dir: 3,
           victory_hexes: [
-            [1, 6, 2], [3, 7, 2], [4, 2, 2], [7, 4, 2], [9, 6, 2], [6, 7, 2], [8, 8, 2],
-            [10, 8, 2], [12, 8, 2],
+            [1, 6, 2], [3, 7, 2], [4, 2, 2], [7, 4, 2], [9, 6, 2], [6, 7, 2], [10, 8, 2],
+            [12, 8, 2], [10, 3, 1],
           ],
           allied_setup: {
             "0" => [
-              ["*", "0-1"], ["5-14", 2], ["7-14", 3], ["8-14", 4], ["10-14", 5],
-              ["11-14", "6-7"], ["13-14", "8-9"],
+              ["*", "0-1"], ["5-14", 2], ["7-14", 3], ["8-14", 4], ["10-14", 5], ["11-14", "6-7"],
+              ["13-14", "8-9"],
             ],
-            "2" => [["6-14", 0]],
+            "2" => [
+              ["6-14", "0-1"], ["12-14", "2-3"], ["13-14", "4-5"], [14, "6-7"],
+            ],
+            "4" => [
+              ["6-14", "0-1"], ["12-14", "2-3"], ["13-14", "4-5"], [14, "6-7"],
+            ],
           },
           axis_setup: {
             "0" => [
@@ -151,7 +156,8 @@ module Scenarios
             { t: "y", r: { d: [3, 6], t: "d" } },
             { t: "y" },
             { t: "y" },
-          ], [
+          ],
+          [
             { t: "s" },
             { t: "s" },
             { t: "s" },
@@ -167,7 +173,8 @@ module Scenarios
             { t: "y" },
             { t: "y" },
             { t: "y" },
-          ], [
+          ],
+          [
             { t: "o", h: 1 },
             { t: "o", h: 1, d: 1, st: { sh: "l2", s: "f" } },
             { t: "o", h: 1, d: 1, st: { sh: "l2", s: "f" } },
@@ -183,7 +190,8 @@ module Scenarios
             { t: "y" },
             { t: "y" },
             { t: "y" },
-          ], [
+          ],
+          [
             { t: "p", h: 1 },
             { t: "p", h: 1 },
             { t: "o", h: 1 },
@@ -199,7 +207,8 @@ module Scenarios
             { t: "y" },
             { t: "y" },
             { t: "y" },
-          ], [
+          ],
+          [
             { t: "p", h: 1 },
             { t: "p", h: 1 },
             { t: "p", h: 1 },
@@ -215,7 +224,8 @@ module Scenarios
             { t: "y" },
             { t: "y" },
             { t: "y" },
-          ], [
+          ],
+          [
             { t: "o", h: 1 },
             { t: "o", h: 1 },
             { t: "o", h: 1 },
@@ -231,7 +241,8 @@ module Scenarios
             { t: "y" },
             { t: "y" },
             { t: "y" },
-          ], [
+          ],
+          [
             { t: "o", h: 1, r: { d: [1, 4], t: "a" } },
             { t: "o", h: 1, r: { d: [1, 4], t: "a" } },
             { t: "o", h: 1, r: { d: [1, 4], t: "a" } },
@@ -247,7 +258,8 @@ module Scenarios
             { t: "y" },
             { t: "y" },
             { t: "y" },
-          ], [
+          ],
+          [
             { t: "o", h: 1 },
             { t: "p", h: 1 },
             { t: "p", h: 1 },
@@ -263,7 +275,8 @@ module Scenarios
             { t: "s" },
             { t: "y" },
             { t: "y" },
-          ], [
+          ],
+          [
             { t: "p", h: 1 },
             { t: "p", h: 1 },
             { t: "p", h: 1 },
@@ -279,7 +292,8 @@ module Scenarios
             { t: "o", h: 1, d: 3.5, st: { sh: "l2", s: "f" } },
             { t: "s" },
             { t: "y" },
-          ], [
+          ],
+          [
             { t: "p", h: 1 },
             { t: "p", h: 1 },
             { t: "p", h: 1 },
@@ -295,7 +309,8 @@ module Scenarios
             { t: "o", h: 1, d: 2, st: { sh: "l2", s: "f" } },
             { t: "s" },
             { t: "s" },
-          ], [
+          ],
+          [
             { t: "o", h: 1 },
             { t: "o", h: 1 },
             { t: "o", h: 1 },